# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Police Audit API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export type IncidentStatus =
  (typeof IncidentStatus)[keyof typeof IncidentStatus];

export const IncidentStatus = {
  recording: "recording",
  pending: "pending",
  analyzed: "analyzed",
  reported: "reported",
} as const;

export interface Incident {
  id: string;
  title: string;
  description?: string;
  location?: string;
  latitude?: number;
  longitude?: number;
  status: IncidentStatus;
  officer_badge?: string;
  officer_name?: string;
  duration_seconds?: number;
  notes?: string;
  created_at: string;
  updated_at?: string;
}

export interface CreateIncidentInput {
  title: string;
  description?: string;
  location?: string;
  latitude?: number;
  longitude?: number;
  officer_badge?: string;
  officer_name?: string;
  duration_seconds?: number;
  notes?: string;
}

export type UpdateIncidentInputStatus =
  (typeof UpdateIncidentInputStatus)[keyof typeof UpdateIncidentInputStatus];

export const UpdateIncidentInputStatus = {
  recording: "recording",
  pending: "pending",
  analyzed: "analyzed",
  reported: "reported",
} as const;

export interface UpdateIncidentInput {
  title?: string;
  description?: string;
  location?: string;
  latitude?: number;
  longitude?: number;
  status?: UpdateIncidentInputStatus;
  officer_badge?: string;
  officer_name?: string;
  duration_seconds?: number;
  notes?: string;
}

export type IncidentEventType =
  (typeof IncidentEventType)[keyof typeof IncidentEventType];

export const IncidentEventType = {
  force_used: "force_used",
  weapon_drawn: "weapon_drawn",
  verbal_aggression: "verbal_aggression",
  unlawful_search: "unlawful_search",
  miranda_rights: "miranda_rights",
  obstruction: "obstruction",
  other: "other",
} as const;

export interface IncidentEvent {
  id: string;
  incident_id: string;
  type: IncidentEventType;
  description?: string;
  timestamp_seconds?: number;
  confidence?: number;
  created_at: string;
}

export type CreateEventInputType =
  (typeof CreateEventInputType)[keyof typeof CreateEventInputType];

export const CreateEventInputType = {
  force_used: "force_used",
  weapon_drawn: "weapon_drawn",
  verbal_aggression: "verbal_aggression",
  unlawful_search: "unlawful_search",
  miranda_rights: "miranda_rights",
  obstruction: "obstruction",
  other: "other",
} as const;

export interface CreateEventInput {
  type: CreateEventInputType;
  description?: string;
  timestamp_seconds?: number;
  confidence?: number;
}

export interface Officer {
  id: string;
  name?: string;
  badge_no?: string;
  agency?: string;
  rank?: string;
  department?: string;
  notes?: string;
  incident_count?: number;
  created_at: string;
}

export interface CreateOfficerInput {
  name?: string;
  badge_no?: string;
  agency?: string;
  rank?: string;
  department?: string;
  notes?: string;
}

export interface Report {
  id: string;
  incident_id: string;
  title?: string;
  summary?: string;
  findings?: string[];
  recommendations?: string[];
  created_at: string;
}

export type ListIncidents200 = {
  incidents: Incident[];
};

export type ListIncidentEvents200 = {
  events: IncidentEvent[];
};

export type ListOfficers200 = {
  officers: Officer[];
};

export type ResolveOfficerBody = {
  badge_number?: string;
  badge_text?: string;
};

export type ResolveOfficer200 = {
  officer?: Officer;
  found: boolean;
  badge_number: string;
};

export type ListReports200 = {
  reports: Report[];
};