# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Police Audit API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: incidents
    description: Incident recording and management
  - name: officers
    description: Officer lookup and records
  - name: reports
    description: PDF report generation
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /incidents:
    get:
      operationId: listIncidents
      tags: [incidents]
      summary: List all incidents
      responses:
        "200":
          description: List of incidents
          content:
            application/json:
              schema:
                type: object
                properties:
                  incidents:
                    type: array
                    items:
                      $ref: "#/components/schemas/Incident"
                required:
                  - incidents

    post:
      operationId: createIncident
      tags: [incidents]
      summary: Create a new incident record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncidentInput"
      responses:
        "201":
          description: Incident created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"

  /incidents/{id}:
    get:
      operationId: getIncident
      tags: [incidents]
      summary: Get a specific incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Incident details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "404":
          description: Incident not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateIncident
      tags: [incidents]
      summary: Update an incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIncidentInput"
      responses:
        "200":
          description: Incident updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"

    delete:
      operationId: deleteIncident
      tags: [incidents]
      summary: Delete an incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Incident deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /incidents/{id}/events:
    get:
      operationId: listIncidentEvents
      tags: [incidents]
      summary: List events for an incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/IncidentEvent"
                required:
                  - events

    post:
      operationId: addIncidentEvent
      tags: [incidents]
      summary: Add an event to an incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventInput"
      responses:
        "201":
          description: Event added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentEvent"

  /officers:
    get:
      operationId: listOfficers
      tags: [officers]
      summary: List all officer records
      responses:
        "200":
          description: List of officers
          content:
            application/json:
              schema:
                type: object
                properties:
                  officers:
                    type: array
                    items:
                      $ref: "#/components/schemas/Officer"
                required:
                  - officers

    post:
      operationId: createOfficer
      tags: [officers]
      summary: Create or save an officer record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOfficerInput"
      responses:
        "201":
          description: Officer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Officer"

  /officers/{id}:
    get:
      operationId: getOfficer
      tags: [officers]
      summary: Get a specific officer record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Officer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Officer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /officers/resolve:
    post:
      operationId: resolveOfficer
      tags: [officers]
      summary: Resolve officer by badge number or OCR text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                badge_number:
                  type: string
                badge_text:
                  type: string
      responses:
        "200":
          description: Officer resolution result
          content:
            application/json:
              schema:
                type: object
                properties:
                  officer:
                    $ref: "#/components/schemas/Officer"
                  found:
                    type: boolean
                  badge_number:
                    type: string
                required:
                  - found
                  - badge_number

  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List all generated reports
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items:
                      $ref: "#/components/schemas/Report"
                required:
                  - reports

  /reports/{incidentId}:
    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate a report for an incident
      parameters:
        - name: incidentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Report generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

    get:
      operationId: getReport
      tags: [reports]
      summary: Get report for an incident
      parameters:
        - name: incidentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    Incident:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        location:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        status:
          type: string
          enum: [recording, pending, analyzed, reported]
        officer_badge:
          type: string
        officer_name:
          type: string
        duration_seconds:
          type: number
        notes:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - title
        - status
        - created_at

    CreateIncidentInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        officer_badge:
          type: string
        officer_name:
          type: string
        duration_seconds:
          type: number
        notes:
          type: string
      required:
        - title

    UpdateIncidentInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        status:
          type: string
          enum: [recording, pending, analyzed, reported]
        officer_badge:
          type: string
        officer_name:
          type: string
        duration_seconds:
          type: number
        notes:
          type: string

    IncidentEvent:
      type: object
      properties:
        id:
          type: string
        incident_id:
          type: string
        type:
          type: string
          enum: [force_used, weapon_drawn, verbal_aggression, unlawful_search, miranda_rights, obstruction, other]
        description:
          type: string
        timestamp_seconds:
          type: number
        confidence:
          type: number
        created_at:
          type: string
      required:
        - id
        - incident_id
        - type
        - created_at

    CreateEventInput:
      type: object
      properties:
        type:
          type: string
          enum: [force_used, weapon_drawn, verbal_aggression, unlawful_search, miranda_rights, obstruction, other]
        description:
          type: string
        timestamp_seconds:
          type: number
        confidence:
          type: number
      required:
        - type

    Officer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        badge_no:
          type: string
        agency:
          type: string
        rank:
          type: string
        department:
          type: string
        notes:
          type: string
        incident_count:
          type: number
        created_at:
          type: string
      required:
        - id
        - created_at

    CreateOfficerInput:
      type: object
      properties:
        name:
          type: string
        badge_no:
          type: string
        agency:
          type: string
        rank:
          type: string
        department:
          type: string
        notes:
          type: string

    Report:
      type: object
      properties:
        id:
          type: string
        incident_id:
          type: string
        title:
          type: string
        summary:
          type: string
        findings:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        created_at:
          type: string
      required:
        - id
        - incident_id
        - created_at