# lib/api-zod/src/generated/types/createEventInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Police Audit API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateEventInputType } from "./createEventInputType";

export interface CreateEventInput {
  type: CreateEventInputType;
  description?: string;
  timestamp_seconds?: number;
  confidence?: number;
}