# lib/api-zod/src/generated/types/createIncidentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Police Audit API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateIncidentInput {
  title: string;
  description?: string;
  location?: string;
  latitude?: number;
  longitude?: number;
  officer_badge?: string;
  officer_name?: string;
  duration_seconds?: number;
  notes?: string;
}