# lib/api-zod/src/generated/types/createOfficerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Police Audit API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOfficerInput {
  name?: string;
  badge_no?: string;
  agency?: string;
  rank?: string;
  department?: string;
  notes?: string;
}