# lib/api-zod/src/generated/types/officer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Police Audit API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Officer {
  id: string;
  name?: string;
  badge_no?: string;
  agency?: string;
  rank?: string;
  department?: string;
  notes?: string;
  incident_count?: number;
  created_at: string;
}