# lib/api-zod/src/generated/types/updateIncidentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Police Audit API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateIncidentInputStatus } from "./updateIncidentInputStatus";

export interface UpdateIncidentInput {
  title?: string;
  description?: string;
  location?: string;
  latitude?: number;
  longitude?: number;
  status?: UpdateIncidentInputStatus;
  officer_badge?: string;
  officer_name?: string;
  duration_seconds?: number;
  notes?: string;
}